Line 531, getMeFollowers()
The getMeFollowers method does not work. I'm not sure about the details of the API so I don't have a proposal but as it is, in attempting to return the ArrayList, it returns a User object from get(). That User object is passed up to the calling getMeFollowers method where a class cast exception is thrown.